+++ weaver/utils.py
@@ -17,6 +17,7 @@ import warnings
 from concurrent.futures import ALL_COMPLETED, CancelledError, ThreadPoolExecutor, as_completed, wait as wait_until
 from copy import deepcopy
 from datetime import datetime
+from pkgutil import get_loader
 from typing import TYPE_CHECKING, overload
 from urllib.parse import ParseResult, unquote, urlparse, urlunsplit
 
@@ -64,13 +65,15 @@ from weaver.warning import TimeZoneInfoAlreadySetWarning
 from weaver.xml_util import HTML_TREE_BUILDER, XML
 
 if TYPE_CHECKING:
-    from types import FrameType
+    import importlib.abc
+    from types import FrameType, ModuleType
     from typing import (
         Any,
         AnyStr,
         Callable,
         Dict,
         List,
+        IO,
         Iterable,
         Iterator,
         MutableMapping,
@@ -988,6 +991,28 @@ def import_target(target, default_root=None):
     mod = importlib.util.module_from_spec(mod_spec)
     mod_spec.loader.exec_module(mod)
     return getattr(mod, target, None)
+
+
+def open_module_resource_file(module, file_path):
+    # type: (Union[str, ModuleType], str) -> IO[bytes]
+    """
+    Opens a resource (data file) from an installed module.
+
+    :returns: File stream handler to read contents as needed.
+    """
+    loader = get_loader(module)
+    # Python <=3.6, no 'get_resource_reader' or 'open_resource' on loader/reader
+    # Python >=3.10, no 'open_resource' directly on loader
+    # Python 3.7-3.9, both permitted in combination
+    try:
+        try:
+            reader = loader.get_resource_reader()  # type: importlib.abc.ResourceReader  # noqa
+        except AttributeError:
+            reader = loader  # noqa
+        return reader.open_resource(file_path)
+    except AttributeError:
+        path = os.path.join(module.__path__[0], file_path)
+        return open(path, mode="r", encoding="utf-8")
 
 
 def now(tz_name=None):
@@ -2917,7 +2942,7 @@ def fetch_reference(reference,                          # type: str
     """
     if reference.endswith("/"):
         path = fetch_directory(reference, out_dir, out_method=out_method, settings=settings, **option_kwargs)
-        path = path if out_listing else (f"{os.path.realpath(out_dir)}/")
+        path = path if out_listing else f"{os.path.realpath(out_dir)}/"
     else:
         path = fetch_file(reference, out_dir, out_method=out_method, settings=settings, **option_kwargs)
     return [path] if out_listing and isinstance(path, str) else path